+++ src/components/ProductsBlock/ProductCard/Buttons/index.tsx
@@ -1,11 +1,15 @@
 import classNames from 'classnames';
+import { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router';
 
 import { ReactComponent as ArrowRight } from '@assets/svg/arrow-right.svg';
 import { ReactComponent as Heart } from '@assets/svg/heart.svg';
+import { Modal } from '@components/Modal';
+import { SingInModal } from '@components/SignInModal';
 import { paths } from '@constants/paths';
 import { useAppDispatch } from '@hooks/useAppDispatch';
 import { useAppSelector } from '@hooks/useAppSelector';
+import { useAuth } from '@hooks/useAuth';
 import {
   addItemToFavorite,
   removeItemFromFavorite,
@@ -22,6 +26,9 @@ export const Buttons: React.FC<PropTypes> = ({ product }) => {
   const dispatch = useAppDispatch();
   const { itemsInFavorite } = useAppSelector((state) => state.wishlist);
 
+  const [isModalActive, setIsModalActive] = useState(false);
+  const { isAuth } = useAuth();
+
   const navigate = useNavigate();
   const location = useLocation();
 
@@ -30,8 +37,15 @@ export const Buttons: React.FC<PropTypes> = ({ product }) => {
   };
 
   const isProductInFavorite = itemsInFavorite.some((el) => el === product.id);
+  const isActiveWishList = isAuth && isProductInFavorite;
 
   const handleAddInWish = () => {
+    if (!isAuth) {
+      setIsModalActive(true);
+
+      return;
+    }
+
     if (isProductInFavorite) {
       dispatch(removeItemFromFavorite(product.id));
     } else {
@@ -59,13 +73,24 @@ export const Buttons: React.FC<PropTypes> = ({ product }) => {
           'product-buttons__button',
           'product-buttons__wishlist',
           {
-            'product-buttons__wishlist--active': isProductInFavorite,
+            'product-buttons__wishlist--active': isActiveWishList,
           },
         )}
       >
         <Heart className="product-buttons__svg product-buttons__svg--heart" />
-        {isProductInFavorite ? 'Remove from wishlist' : 'Add to wishlist'}
+        {isActiveWishList ? 'In wishlist' : 'Add to wish list'}
       </button>
+      {isModalActive && (
+        <Modal
+          setIsModalActive={setIsModalActive}
+          isModalActive={isModalActive}
+        >
+          <SingInModal
+            setIsModalActive={setIsModalActive}
+            id={product.id}
+          />
+        </Modal>
+      )}
     </div>
   );
 };
